+++ src/Knp/Bundle/KnpBundlesBundle/Updater/Exception/UserNotFoundException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Knp\Bundle\KnpBundlesBundle\Updater\Exception;
-
-class UserNotFoundException extends \Exception
-{
-
-}